Move PathBuilder and PathCleaner under Path::
