+++ spec/draught/path/cleaner_spec.rb
@@ -1,8 +1,8 @@
-require 'draught/path_cleaner'
+require 'draught/path/cleaner'
 require 'draught/world'
 
 module Draught
-  RSpec.describe PathCleaner do
+  RSpec.describe Path::Cleaner do
     let(:world) { World.new }
     let(:metadata) { Metadata::Instance.new(name: 'name') }
 
@@ -15,12 +15,12 @@ module Draught
         input = world.path.new(points: [p(0,0), p(1,0), p(1,0), p(4,0)])
         expected = world.path.new(points: [p(0,0), p(1,0), p(4,0)])
 
-        expect(PathCleaner.dedupe(world, input)).to eq(expected)
+        expect(described_class.dedupe(world, input)).to eq(expected)
       end
 
       context "handling Metadata" do
         let(:input) { world.path.new(points: [p(0,0), p(4,0), p(1,0)], metadata: metadata) }
-        let(:deduped) { PathCleaner.dedupe(world, input) }
+        let(:deduped) { described_class.dedupe(world, input) }
 
         specify "preserves Metadata" do
           expect(deduped.metadata).to be(metadata)
@@ -34,52 +34,52 @@ module Draught
       it "removes unnecessary intermediate points on the horizontal" do
         input = world.path.new(points: [p(0,0), p(1,0), p(1,0), p(2,0), p(3,0), p(4,0), p(4,4)])
 
-        expect(PathCleaner.simplify(world, input)).to eq(expected)
+        expect(described_class.simplify(world, input)).to eq(expected)
       end
 
       it "removes unnecessary intermediate points on the horizontal" do
         input = world.path.new(points: [p(0,0), p(1,0), p(1,0), p(2,0), p(3,0), p(4,0), p(4,4)])
 
-        expect(PathCleaner.simplify(world, input)).to eq(expected)
+        expect(described_class.simplify(world, input)).to eq(expected)
       end
 
       it "removes unnecessary intermediate points on the vertical" do
         input = world.path.new(points: [p(0,0), p(4,0), p(4,1), p(4,2), p(4,4)])
 
-        expect(PathCleaner.simplify(world, input)).to eq(expected)
+        expect(described_class.simplify(world, input)).to eq(expected)
       end
 
       it "goes around corners" do
         input = world.path.new(points: [p(0,0), p(1,0), p(3,0), p(4,0), p(4,2), p(4,4)])
 
-        expect(PathCleaner.simplify(world, input)).to eq(expected)
+        expect(described_class.simplify(world, input)).to eq(expected)
       end
 
       it "does not remove points where it's not obvious which point to remove" do
         input = world.path.new(points: [p(0,0), p(4,0), p(1,0)])
         expected = world.path.new(points: [p(0,0), p(4,0), p(1,0)])
 
-        expect(PathCleaner.simplify(world, input)).to eq(expected)
+        expect(described_class.simplify(world, input)).to eq(expected)
       end
 
       it "copes with paths proceeding right-to-left" do
         input = world.path.new(points: [p(4,0), p(2,0), p(1,0)])
         expected = world.path.new(points: [p(4,0), p(1,0)])
 
-        expect(PathCleaner.simplify(world, input)).to eq(expected)
+        expect(described_class.simplify(world, input)).to eq(expected)
       end
 
       it "copes with paths proceeding top-to-bottom" do
         input = world.path.new(points: [p(0,4), p(0,2), p(0,1)])
         expected = world.path.new(points: [p(0,4), p(0,1)])
 
-        expect(PathCleaner.simplify(world, input)).to eq(expected)
+        expect(described_class.simplify(world, input)).to eq(expected)
       end
 
 
       context "handling Metadata" do
         let(:input) { world.path.new(points: [p(0,0), p(4,0), p(1,0)], metadata: metadata) }
-        let(:deduped) { PathCleaner.simplify(world, input) }
+        let(:deduped) { described_class.simplify(world, input) }
 
         specify "preserves Metadata" do
           expect(deduped.metadata).to be(metadata)
